+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Hypergraph.edge_count().
 - Hypergraph.primary_atom_count().
 - Hypergraph.primary_edge_count().
+- info command.
 
 ## [0.4.3] - 22-04-2021
 ### Changed