+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [0.5.0] - ongoing work
+## [0.5.0] - 28-07-2021
 ### Added
 - SQLite3 hypergraph database backend.
 - Hypergraph.add_with_attributes().